Trait pallet_msa_rpc::MsaApiClient

pub trait MsaApiClient<BlockHash, AccountId>: ClientTwhere
    BlockHash: Send + Sync + 'static,
    AccountId: Send + Sync + 'static,{
    fn check_delegations<'life0, 'async_trait>(
        &'life0 self,
        delegator_msa_ids: Vec<MessageSourceId>,
        provider_msa_id: MessageSourceId,
        block_number: Option<BlockNumber>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<(MessageSourceId, bool)>>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn get_granted_schemas_by_msa_id<'life0, 'async_trait>(
        &'life0 self,
        delegator_msa_id: MessageSourceId,
        provider_msa_id: MessageSourceId
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<SchemaId>>>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Client implementation for the MsaApi RPC API.

Provided Methods

fn check_delegations<'life0, 'async_trait>(
&'life0 self,
delegator_msa_ids: Vec<MessageSourceId>,
provider_msa_id: MessageSourceId,
block_number: Option<BlockNumber>
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<(MessageSourceId, bool)>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Check for a list of delegations Given a single provider, test a list of potential delegators At a given block number

fn get_granted_schemas_by_msa_id<'life0, 'async_trait>(
&'life0 self,
delegator_msa_id: MessageSourceId,
provider_msa_id: MessageSourceId
) -> Pin<Box<dyn Future<Output = RpcResult<Option<Vec<SchemaId>>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Retrieve the list of currently granted schemas given a delegator and provider pair