Trait pallet_msa_rpc::MsaApiServer

pub trait MsaApiServer<BlockHash, AccountId>: Sized + Send + Sync + 'static {
    fn get_msa_id(&self, key: AccountId) -> RpcResult<Option<MessageSourceId>>;
    fn check_delegations(
        &self,
        delegator_msa_ids: Vec<MessageSourceId>,
        provider_msa_id: MessageSourceId,
        block_number: Option<BlockNumber>
    ) -> RpcResult<Vec<(MessageSourceId, bool)>>;
    fn get_granted_schemas(
        &self,
        delegator_msa_id: MessageSourceId,
        provider_msa_id: MessageSourceId
    ) -> RpcResult<Option<Vec<SchemaId>>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        BlockHash: Send + Sync + 'static,
        AccountId: Send + Sync + 'static + DeserializeOwned,
    { ... }
}

Server trait implementation for the MsaApi RPC API.

Required Methods

fn get_msa_id(&self, key: AccountId) -> RpcResult<Option<MessageSourceId>>

fn check_delegations(
&self,
delegator_msa_ids: Vec<MessageSourceId>,
provider_msa_id: MessageSourceId,
block_number: Option<BlockNumber>
) -> RpcResult<Vec<(MessageSourceId, bool)>>

fn get_granted_schemas(
&self,
delegator_msa_id: MessageSourceId,
provider_msa_id: MessageSourceId
) -> RpcResult<Option<Vec<SchemaId>>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self> where
BlockHash: Send + Sync + 'static,
AccountId: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.